requirements.txt
certifi==2019.6.16
chardet==3.0.4
clearbit==0.1.7
configparser==3.8.1
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.0.0
idna==2.8
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
psycopg2
gunicorn==19.9.0
